fixing norm

+++ converter/ptr.c
@@ -6,7 +6,7 @@
 /*   By: yoav <yoav@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/19 09:57:43 by yoav              #+#    #+#             */
-/*   Updated: 2022/06/19 16:54:12 by yoav             ###   ########.fr       */
+/*   Updated: 2022/06/19 18:50:15 by yoav             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,5 @@ void	convert_ptr(va_list l, void **res, t_flags *f)
 
 	p = va_arg(l, void *);
 	f->hash = TRUE;
-	convert_generic_unsigned_number(p, res, f, BASE_HAX_LOWER);
+	convert_generic_unsigned_number((size_t)p, res, f, BASE_HAX_LOWER);
 }
